# generator/out/backend/openapi/schema.yaml
openapi: 3.0.3
info:
  title: " API"
  version: "1.0.0"
paths:
...
  /beiratkozas/{id}:
    get:
      summary: Get beiratkozas by id
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/Beiratkozas" } } } }
        "404": { description: not found }
    patch:
      summary: Update beiratkozas
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: ok }
        "404": { description: not found }
    delete:
      summary: Delete beiratkozas
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "204": { description: deleted }
        "404": { description: not found }
...
  /hallgato/{id}:
    get:
      summary: Get hallgato by id
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/Hallgato" } } } }
        "404": { description: not found }
    patch:
      summary: Update hallgato
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: ok }
        "404": { description: not found }
    delete:
      summary: Delete hallgato
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "204": { description: deleted }
        "404": { description: not found }
...
components:
  schemas:
    Beiratkozas:
      type: object
      properties:
        hallgato_id:
          type: string
          description: "no comment"
        kurzus_id:
          type: string
          description: "no comment"
        felvet_datum:
          type: string
          description: "no comment"
        jegy:
          type: string
          description: "no comment"
    Hallgato:
      type: object
      properties:
        id:
          type: string
          description: "no comment"
        neptun:
          type: string
          description: "no comment"
        nev:
          type: string
          description: "no comment"
        nem:
          type: string
          description: "no comment"
        szak:
          type: string
          description: "no comment"
        evfolyam:
          type: string
          description: "no comment"
        szuldatum:
          type: string
          description: "no comment"
        email:
          type: string
          description: "no comment"
...